+++ desktop/cores/sdl/Klangstrom/NodeVCA.hpp
@@ -10,7 +10,7 @@
  *       +---------------------+
  *       |                     |
  * IN00--| SIGNAL       SIGNAL |--OUT00
- *       |                     |
+ * IN01--| AMPLIFICATION       |
  *       +---------------------+
  */
 
@@ -20,8 +20,11 @@
 #include "NodeVCA.hpp"
 
 namespace klang {
-    class NodeVCA : public NodeKernel {
+    class NodeVCA : public NodeKernelBlock2 {
     public:
+        static const CHANNEL_ID CH_IN_SIGNAL        = 0;
+        static const CHANNEL_ID CH_IN_AMPLIFICATION = 1;
+
         void set_amplification(SIGNAL_TYPE pValue) {
             mAmplification = pValue;
         }
@@ -39,8 +42,22 @@ namespace klang {
         }
 
     protected:
-        SIGNAL_TYPE kernel(const SIGNAL_TYPE s) {
-            return s * mAmplification;
+        void kernel(SIGNAL_TYPE* pOutputSignal,
+                    SIGNAL_TYPE* pInputSignal_A,
+                    SIGNAL_TYPE* pInputSignal_B) {
+            if (pInputSignal_A == nullptr) {
+                for (uint16_t i = 0; i < KLANG_SAMPLES_PER_AUDIO_BLOCK; i++) {
+                    pOutputSignal[i] = 0.0;
+                }
+            } else if (pInputSignal_B == nullptr) {
+                for (uint16_t i = 0; i < KLANG_SAMPLES_PER_AUDIO_BLOCK; i++) {
+                    pOutputSignal[i] = pInputSignal_A[i] * mAmplification;
+                }
+            } else {
+                for (uint16_t i = 0; i < KLANG_SAMPLES_PER_AUDIO_BLOCK; i++) {
+                    pOutputSignal[i] = pInputSignal_A[i] * pInputSignal_B[i];
+                }
+            }
         }
 
     private: